requirements/base.txt
Django==1.6
PIL==1.1.7
Pygments==1.6
South==0.8.2
beautifulsoup4==4.3.2
bpython==0.12
django-braces==1.2.2
django-mathfilters==0.1.3
django-model-utils==1.4.0
django-suit==0.2.5
django-summernote==0.4.2
django-tagging==0.3.1
logutils==0.3.3
python-dateutil==2.1
six==1.4.1
slugify==0.0.1
stevedore==0.12
virtualenv==1.10.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wsgiref==0.1.2